Simplify component features by only guarding for `codecs`
