+++ Cargo.toml
@@ -448,7 +448,7 @@ sources-docker_logs = ["docker"]
 sources-eventstoredb_metrics = []
 sources-exec = []
 sources-file = ["bytesize", "file-source"]
-sources-fluent = ["base64", "bytesize", "listenfd", "tokio-util/net", "rmpv", "rmp-serde", "sources-utils-tcp-keepalive", "sources-utils-tcp-socket", "sources-utils-tls", "serde_bytes", "codecs"]
+sources-fluent = ["base64", "bytesize", "listenfd", "tokio-util/net", "rmpv", "rmp-serde", "sources-utils-tcp-keepalive", "sources-utils-tcp-socket", "sources-utils-tls", "serde_bytes"]
 sources-generator = ["fakedata"]
 sources-heroku_logs = ["sources-utils-http"]
 sources-host_metrics = ["heim"]
@@ -458,7 +458,7 @@ sources-internal_metrics = []
 sources-journald = ["codecs"]
 sources-kafka = ["rdkafka"]
 sources-nats = ["async-nats"]
-sources-logstash = ["bytesize", "listenfd", "tokio-util/net", "sources-utils-tcp-keepalive", "sources-utils-tcp-socket", "sources-utils-tls", "codecs"]
+sources-logstash = ["bytesize", "listenfd", "tokio-util/net", "sources-utils-tcp-keepalive", "sources-utils-tcp-socket", "sources-utils-tls"]
 sources-kubernetes_logs = ["file-source", "kubernetes", "transforms-merge", "transforms-regex_parser"]
 sources-mongodb_metrics = ["mongodb"]
 sources-nginx_metrics = ["nom"]
@@ -475,7 +475,7 @@ sources-utils-tcp-socket = []
 sources-utils-tls = []
 sources-utils-udp = []
 sources-utils-unix = []
-sources-vector = ["listenfd", "sources-utils-tcp-keepalive", "sources-utils-tcp-socket", "sources-utils-tls", "tonic", "tonic-build", "prost-build", "codecs"]
+sources-vector = ["listenfd", "sources-utils-tcp-keepalive", "sources-utils-tcp-socket", "sources-utils-tls", "tonic", "tonic-build", "prost-build"]
 
 # Transforms
 transforms = ["transforms-logs", "transforms-metrics"]
@@ -628,18 +628,16 @@ sinks-logdna = ["bytesize"]
 sinks-loki = ["bytesize", "uuid"]
 sinks-nats = ["async-nats"]
 sinks-new_relic_logs = ["bytesize", "sinks-http"]
-sinks-papertrail = ["syslog", "sinks-utils-tcp"]
+sinks-papertrail = ["syslog"]
 sinks-prometheus = ["prometheus-parser", "snap", "sources-utils-tls"]
 sinks-pulsar = ["avro-rs", "pulsar"]
 sinks-redis = ["redis"]
 sinks-sematext = ["sinks-elasticsearch", "sinks-influxdb"]
-sinks-socket = ["sinks-utils-tcp", "sinks-utils-udp", "sinks-utils-unix"]
+sinks-socket = ["sinks-utils-udp"]
 sinks-splunk_hec = ["bytesize"]
-sinks-statsd = ["sinks-utils-tcp", "sinks-utils-udp", "tokio-util/net"]
-sinks-utils-tcp = []
+sinks-statsd = ["sinks-utils-udp", "tokio-util/net"]
 sinks-utils-udp = []
-sinks-utils-unix = []
-sinks-vector = ["sinks-utils-tcp", "sinks-utils-udp", "tonic", "tonic-build", "prost-build"]
+sinks-vector = ["sinks-utils-udp", "tonic", "tonic-build", "prost-build"]
 
 # Datadog integration
 datadog-pipelines = ["sources-internal_metrics", "sinks-datadog"]
