+++ src/sinks/util/socket_bytes_sink.rs
@@ -1,4 +1,7 @@
-use crate::{buffers::Acker, internal_events::SocketEventsSent, socket::SocketMode};
+use crate::{
+    buffers::Acker,
+    internal_events::{SocketEventsSent, SocketMode},
+};
 use bytes::Bytes;
 use futures::{ready, Sink};
 use pin_project::{pin_project, pinned_drop};
