+++ src/sinks/util/udp.rs
@@ -6,14 +6,13 @@ use crate::{
     dns,
     event::Event,
     internal_events::{
-        SocketEventsSent, UdpSendIncomplete, UdpSocketConnectionEstablished,
+        SocketEventsSent, SocketMode, UdpSendIncomplete, UdpSocketConnectionEstablished,
         UdpSocketConnectionFailed, UdpSocketError,
     },
     sinks::{
         util::{retries::ExponentialBackoff, StreamSink},
         Healthcheck, VectorSink,
     },
-    socket::SocketMode,
 };
 use async_trait::async_trait;
 use bytes::Bytes;
