+++ src/sources/socket/unix.rs
@@ -1,9 +1,8 @@
 use crate::{
     codecs::Decoder,
     event::Event,
-    internal_events::SocketEventsReceived,
+    internal_events::{SocketEventsReceived, SocketMode},
     shutdown::ShutdownSignal,
-    socket::SocketMode,
     sources::{
         util::{build_unix_datagram_source, build_unix_stream_source},
         Source,
